requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
python-dateutil
openpyxl
pytest
